+++ extension/lib/addProductsToCart.js
@@ -2,6 +2,7 @@
 
 const ShopgateCartPipeline = require('./shopgate/CartExtensionPipeline')
 const { decorateError } = require('./shopgate/logDecorator')
+const ShopgateCartMessage = require('./shopgate/cart/Message')
 
 /**
  * @param {PipelineContext} context
@@ -24,10 +25,17 @@ module.exports = async (context, input) => {
 
   let shopgateCartPipeline = ShopgateCartPipeline.create(context)
   try {
-    await shopgateCartPipeline.addProducts(input.products)
+    return await shopgateCartPipeline.addProducts(input.products).catch((err) => {
+      if (err.code === 422) {
+        const message = new ShopgateCartMessage('error', 'This product is not available anymore', 422)
+        return {
+          messages: [message.toJson()]
+        }
+      }
+      return {}
+    })
   } catch (error) {
     context.log.error(decorateError(error), 'Failed adding products to cart')
     throw new Error()
   }
-  return {}
 }